refactor(credentials): Use CredentialsNavigator

+++ kork-credentials/src/main/java/com/netflix/spinnaker/credentials/service/CredentialsDefinitionService.java
@@ -24,9 +24,11 @@ import com.netflix.spinnaker.credentials.validator.CredentialsDefinitionErrorCod
 import com.netflix.spinnaker.credentials.validator.CredentialsDefinitionValidator;
 import com.netflix.spinnaker.kork.annotations.NonnullByDefault;
 import com.netflix.spinnaker.kork.web.exceptions.ValidationException;
+import com.netflix.spinnaker.security.AccessControlled;
 import com.netflix.spinnaker.security.Authorization;
 import java.util.Collection;
 import java.util.List;
+import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
@@ -34,8 +36,6 @@ import lombok.extern.log4j.Log4j2;
 import org.springframework.beans.factory.ObjectProvider;
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.access.PermissionEvaluator;
-import org.springframework.security.access.prepost.PostFilter;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import org.springframework.validation.BeanPropertyBindingResult;
@@ -60,6 +60,25 @@ public class CredentialsDefinitionService {
         || credentialsRepositories.stream().anyMatch(repo -> repo.has(name));
   }
 
+  private boolean shouldDenyWritePermission(Authentication auth, String name) {
+    return repository
+        .findByName(name)
+        .filter(AccessControlled.class::isInstance)
+        .map(AccessControlled.class::cast)
+        .or(
+            () ->
+                credentialsRepositories
+                    .orderedStream()
+                    .filter(repo -> repo.has(name))
+                    .map(repo -> repo.getOne(name))
+                    .filter(Objects::nonNull)
+                    .filter(AccessControlled.class::isInstance)
+                    .map(AccessControlled.class::cast)
+                    .findFirst())
+        .map(existing -> !permissionEvaluator.hasPermission(auth, existing, Authorization.WRITE))
+        .orElse(false);
+  }
+
   /**
    * Validates the given credential definition for the provided user for the given command.
    *
@@ -70,7 +89,6 @@ public class CredentialsDefinitionService {
    * @throws AccessDeniedException if the user is unauthorized to perform the given command
    * @see CredentialsDefinitionValidator
    */
-  @PreAuthorize("hasAnyAuthority('SPINNAKER_ADMIN', 'SPINNAKER_ACCOUNT_MANAGER')")
   public void validate(
       CredentialsDefinition definition, Authentication auth, CredentialsDefinitionCommand command) {
     String name = definition.getName();
@@ -89,7 +107,7 @@ public class CredentialsDefinitionService {
             CredentialsDefinitionErrorCode.DUPLICATE_NAME.getErrorCode(),
             "Cannot create a new account with the same name as an existing one");
       }
-      if (!permissionEvaluator.hasPermission(auth, name, "account", Authorization.WRITE)) {
+      if (shouldDenyWritePermission(auth, name)) {
         throw new AccessDeniedException("Unauthorized to overwrite existing account");
       }
     }
@@ -99,26 +117,22 @@ public class CredentialsDefinitionService {
     }
   }
 
-  @PreAuthorize("hasAnyAuthority('SPINNAKER_ADMIN', 'SPINNAKER_ACCOUNT_MANAGER')")
   public void create(CredentialsDefinition definition, Authentication auth) {
     validate(definition, auth, CredentialsDefinitionCommand.CREATE);
     repository.create(definition);
   }
 
-  @PreAuthorize("hasAnyAuthority('SPINNAKER_ADMIN', 'SPINNAKER_ACCOUNT_MANAGER')")
   public void save(CredentialsDefinition definition, Authentication auth) {
     validate(definition, auth, CredentialsDefinitionCommand.SAVE);
     repository.save(definition);
   }
 
-  @PreAuthorize("hasAnyAuthority('SPINNAKER_ADMIN', 'SPINNAKER_ACCOUNT_MANAGER')")
   public void saveAll(Collection<CredentialsDefinition> definitions, Authentication auth) {
     var errors = new BeanPropertyBindingResult(definitions, "definitions");
     for (CredentialsDefinition definition : definitions) {
       String name = definition.getName();
       errors.pushNestedPath(name);
-      if (isNameInUse(name)
-          && !permissionEvaluator.hasPermission(auth, name, "account", Authorization.WRITE)) {
+      if (shouldDenyWritePermission(auth, name)) {
         errors.rejectValue(
             "name",
             CredentialsDefinitionErrorCode.UNAUTHORIZED.getErrorCode(),
@@ -133,25 +147,23 @@ public class CredentialsDefinitionService {
     repository.saveAll(definitions);
   }
 
-  @PreAuthorize("hasAnyAuthority('SPINNAKER_ADMIN', 'SPINNAKER_ACCOUNT_MANAGER')")
   public void update(CredentialsDefinition definition, Authentication auth) {
     validate(definition, auth, CredentialsDefinitionCommand.UPDATE);
     repository.update(definition);
   }
 
-  @PreAuthorize(
-      "hasAuthority('SPINNAKER_ADMIN') or hasAuthority('SPINNAKER_ACCOUNT_MANAGER') and hasPermission(#accountName, 'account', 'write')")
-  public void delete(String accountName) {
+  public void delete(String accountName, Authentication auth) {
+    if (shouldDenyWritePermission(auth, accountName)) {
+      throw new AccessDeniedException(
+          String.format("Unauthorized to delete account '%s'", accountName));
+    }
     repository.delete(accountName);
   }
 
-  @PreAuthorize("hasAnyAuthority('SPINNAKER_ADMIN', 'SPINNAKER_ACCOUNT_MANAGER')")
   public void deleteAll(Collection<String> accountNames, Authentication auth) {
     Set<String> unauthorizedAccountNamesToDelete =
         accountNames.stream()
-            .filter(
-                name ->
-                    !permissionEvaluator.hasPermission(auth, name, "account", Authorization.WRITE))
+            .filter(name -> shouldDenyWritePermission(auth, name))
             .collect(Collectors.toSet());
     if (!unauthorizedAccountNamesToDelete.isEmpty()) {
       throw new AccessDeniedException(
@@ -160,8 +172,6 @@ public class CredentialsDefinitionService {
     repository.deleteAll(accountNames);
   }
 
-  @PreAuthorize("hasAnyAuthority('SPINNAKER_ADMIN', 'SPINNAKER_ACCOUNT_MANAGER')")
-  @PostFilter("hasPermission(filterObject.account, 'write')")
   public List<CredentialsDefinitionRepository.Revision> revisionHistory(String accountName) {
     return repository.revisionHistory(accountName);
   }
